requirements.txt
# requirements.txt

# UI
streamlit

# Existing
chromadb
sentence-transformers
# tqdm

# LLM Providers
openai
groq
google-genai

# Utils
python-dotenv
pydantic-settings